requirements.txt
# requirements.txt
# Gerekli Python paketleri
# Bu dosya, uygulamanın çalışması için gerekli tüm bağımlılıkları içerir

# Flask ve ilgili uzantılar
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-Admin==1.6.1
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6  # Yeni eklenen
google-generativeai==0.5.2
librosa==0.10.1
scipy==1.13.1
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
pydub==0.25.1
celery==5.3.6
redis==5.0.4
python-dotenv==1.0.1
email-validator==2.1.1
gunicorn==22.0.0
eventlet==0.33.3

# Opsiyonel AI Providers (ihtiyaca göre yükleyin)
# openai>=1.0.0              # OpenAI GPT-4 desteği için
# anthropic>=0.18.0          # Anthropic Claude desteği için  # Yeni eklenen

# Geliştirme araçları
pytest==8.2.2
pytest-cov==5.0.0
flake8==7.1.0
black==24.4.2